TDD/TDD_With_JUnit5_And_Mockito/code/1-implementing-tdd-and-code-coverage/flight-mgmt-app/flight-mgmt-app/src/test/java/com/shshetudev/flightmgmtapp/airport/FlightTest.java
package com.shshetudev.flightmgmtapp.airport;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class FlightTest {

    @Test
    void testFlight() {
        Flight economy = new Flight("1", "Economy");
        Flight business = new Flight("2", "Business");

        Passenger john = new Passenger("John", true);
        Passenger mike = new Passenger("Mike", false);
        // vip
        assertTrue(economy.addPassenger(john));
        assertFalse(economy.removePassenger(john));

        assertTrue(business.addPassenger(john));
        assertFalse(business.removePassenger(john));

        //non-vip
        assertTrue(economy.addPassenger(mike));
        assertTrue(economy.removePassenger(mike));

        assertFalse(business.addPassenger(mike));
        assertFalse(business.removePassenger(mike));
    }
}
